+++ src/main/java/net/kenvanhoeylandt/solutions/day7/logic/gates/AndGate.java
@@ -10,7 +10,7 @@ public class AndGate extends DoubleInputGate
 	}
 
 	@Override
-	public long getValue()
+	public int getValue()
 	{
 		return getFirstInputValue() & getSecondInputValue();
 	}
