+++ src/main/java/net/kenvanhoeylandt/solutions/day7/logic/valueproviders/GateValueProvider.java
@@ -17,7 +17,7 @@ public class GateValueProvider implements ValueProvider
 	}
 
 	@Override
-	public long getValue()
+	public int getValue()
 	{
 		Gate gate = mGateManager.getGate(mGateName);
